Ajuste para aperfeiçoar rotina de redirecionamento e autenticação localmente

+++ www/js/controllers/ContaCtrl.js
@@ -1,36 +1,54 @@
-app.controller('ContaCtrl', function ($scope, $state, $stateParams, $ionicHistory, utilAPI, usuarioAPILocal, md5) {
+app.controller('ContaCtrl', function ($scope, $state, $stateParams, $ionicHistory, utilAPI, usuarioAPILocal, md5, $ionicSideMenuDelegate) {
     $scope.conta = [];
-    
+
     $scope.$on('$ionicView.beforeEnter', function () {
+        if (usuarioLogin && usuarioLogin.id) {
+            $scope.conta = usuarioLogin;
+            $scope.conta.senha = null;
+            $scope.conta.confirmar = null;
+        } else if ($stateParams.email) {
+            $scope.conta.email = $stateParams.email;
+        }
     });
-    
-    if ($stateParams.email) {
-        $scope.conta.email = $stateParams.email;
-    }
 
     $scope.salvar = function (formValido) {
-        if(formValido){
+        if (formValido) {
             var conta = $scope.conta;
             if (conta && conta.email && conta.senha && conta.confirmar) {
                 if (conta.senha == conta.confirmar) {
                     //Implementar rotina para salvar dados no servidor. Hoje esta apenas salvando a conta localmente.
-                    usuarioAPILocal.delete().then(function () {
+                    if (!Number(conta.id)) {
+                        usuarioAPILocal.delete().then(function () {
+                            var usuario = [];
+                            usuario.nome = conta.nome;
+                            usuario.email = conta.email;
+                            usuario.senha = md5.createHash(conta.senha);
+                            usuarioAPILocal.insert(usuario).then();
+                            usuarioAPILocal.getByEmailSenha(usuario.email, usuario.senha).then(function (res) {
+                                if (res) {
+                                    usuarioLogin = res;
+                                    utilAPI.avisoTemp("Sua conta foi criada");
+                                    $ionicHistory.nextViewOptions({
+                                        disableBack: true
+                                    });
+                                    $state.go("app.projetos");
+                                }
+                            });
+                        });
+                    } else {
                         var usuario = [];
+                        usuario.id = conta.id;
                         usuario.nome = conta.nome;
                         usuario.email = conta.email;
                         usuario.senha = md5.createHash(conta.senha);
-                        usuarioAPILocal.insert(usuario);
-                        usuarioAPILocal.getByEmailSenha(usuario.email, usuario.senha).then(function (res) {
-                            if (res) {
-                                usuarioLogin = res;
-                                utilAPI.avisoTemp("Sua conta foi criada com sucesso!")
-                                $ionicHistory.nextViewOptions({
-                                    disableBack: true
-                                });
-                                $state.go("app.projetos");
-                            }
+                        usuarioAPILocal.edit(usuario).then(function () {
+                            utilAPI.avisoTemp("Sua conta foi atualizada");
+                            $ionicHistory.nextViewOptions({
+                                disableBack: true
+                            });
+                            $state.go("app.projetos");
                         });
-                    });
+                    }
                 } else {
                     utilAPI.avisoTemp("Senhas informadas não coincidem")
                 }
